requirements.txt
# ============================================
# Core Framework
# ============================================
fastapi==0.109.0
uvicorn[standard]==0.27.0

# ============================================
# AI Agent
# ============================================
pydantic-ai>=0.1.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ============================================
# LLM Provider
# ============================================
openai>=1.0.0

# ============================================
# API Integrations
# ============================================
httpx>=0.26.0        # Async HTTP client
pyairtable==2.2.0    # Airtable Python SDK

# ============================================
# Background Processing
# ============================================
redis>=5.0.1
celery>=5.3.0

# ============================================
# Configuration
# ============================================
python-dotenv>=1.0.0
pyyaml>=6.0.1

# ============================================
# Logging
# ============================================
loguru>=0.7.0

# ============================================
# Async Utilities
# ============================================
aiofiles>=23.0.0

# ============================================
# Development & Testing
# ============================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.0.0
ruff>=0.1.0
